backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
djangorestframework==3.15.2
drf-jwt==1.19.2
gunicorn==22.0.0
idna==3.6
invoke==2.2.0
more-itertools==10.2.0
packaging==24.1
psycopg==3.2.1
psycopg-binary==3.2.1
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
whitenoise==6.7.0
